requirement.txt
python=3.8
Pillow=9.5.0
alibabacloud_ocr_api20210707==1.1.15
numpy-1.24.3 
opencv-python==4.7.0.72
pyPDF2-3.0.1
pdf2image-1.16.3(sudo apt-get install poppler-utils)

pytesseract-0.3.10(sudo apt-get install tesseract-ocr)
pdfplumber-0.9.0

paddlepaddle==2.4.2
paddleocr==2.6.1.3